+++ meteor/server/migration/1_47_0.ts
@@ -188,121 +188,121 @@ export const addSteps = addMigrationSteps('1.47.0', [
 		},
 	},
 
-	// {
-	// 	id: `Studios remove pre *WithOverrides properties`,
-	// 	canBeRunAutomatically: true,
-	// 	validate: () => {
-	// 		const objects = Studios.find({
-	// 			$or: [{ blueprintConfig: { $exists: true } }, { mappings: { $exists: true } }],
-	// 		}).count()
-	// 		if (objects > 0) {
-	// 			return `object needs to be converted`
-	// 		}
-	// 		return false
-	// 	},
-	// 	migrate: () => {
-	// 		const objects = Studios.find({
-	// 			$or: [{ blueprintConfig: { $exists: true } }, { mappings: { $exists: true } }],
-	// 		}).fetch()
-	// 		for (const obj0 of objects) {
-	// 			const obj = obj0 as unknown as DBStudio & StudioOld
-	// 			Studios.update(obj._id, {
-	// 				$unset: {
-	// 					blueprintConfig: 1,
-	// 					mappings: 1,
-	// 				},
-	// 			})
-	// 		}
-	// 	},
-	// },
-	// {
-	// 	id: `ShowStyleVariants remove pre *withOverrides properties`,
-	// 	canBeRunAutomatically: true,
-	// 	validate: () => {
-	// 		const objects = ShowStyleVariants.find({
-	// 			blueprintConfig: { $exists: true },
-	// 		}).count()
-	// 		if (objects > 0) {
-	// 			return `object needs to be converted`
-	// 		}
-	// 		return false
-	// 	},
-	// 	migrate: () => {
-	// 		const objects = ShowStyleVariants.find({
-	// 			blueprintConfig: { $exists: true },
-	// 		}).fetch()
-	// 		for (const obj0 of objects) {
-	// 			const obj = obj0 as unknown as DBShowStyleVariant & ShowStyleVariantOld
-	// 			ShowStyleVariants.update(obj._id, {
-	// 				$unset: {
-	// 					blueprintConfig: 1,
-	// 				},
-	// 			})
-	// 		}
-	// 	},
-	// },
-	// {
-	// 	id: `ShowStyleBases remove pre *withOverrides properties`,
-	// 	canBeRunAutomatically: true,
-	// 	validate: () => {
-	// 		const objects = ShowStyleBases.find({
-	// 			$or: [
-	// 				{ blueprintConfig: { $exists: true } },
-	// 				{ sourceLayers: { $exists: true } },
-	// 				{ outputLayers: { $exists: true } },
-	// 			],
-	// 		}).count()
-	// 		if (objects > 0) {
-	// 			return `object needs to be converted`
-	// 		}
-	// 		return false
-	// 	},
-	// 	migrate: () => {
-	// 		const objects = ShowStyleBases.find({
-	// 			$or: [
-	// 				{ blueprintConfig: { $exists: true } },
-	// 				{ sourceLayers: { $exists: true } },
-	// 				{ outputLayers: { $exists: true } },
-	// 			],
-	// 		}).fetch()
-	// 		for (const obj0 of objects) {
-	// 			const obj = obj0 as unknown as DBShowStyleBase & ShowStyleBaseOld
-	// 			ShowStyleBases.update(obj._id, {
-	// 				$unset: {
-	// 					blueprintConfig: 1,
-	// 					sourceLayers: 1,
-	// 					outputLayers: 1,
-	// 				},
-	// 			})
-	// 		}
-	// 	},
-	// },
-	// {
-	// 	id: `TriggeredActions remove pre *withOverrides properties`,
-	// 	canBeRunAutomatically: true,
-	// 	validate: () => {
-	// 		const objects = TriggeredActions.find({
-	// 			$or: [{ triggers: { $exists: true } }, { actions: { $exists: true } }],
-	// 		}).count()
-	// 		if (objects > 0) {
-	// 			return `object needs to be converted`
-	// 		}
-	// 		return false
-	// 	},
-	// 	migrate: () => {
-	// 		const objects = TriggeredActions.find({
-	// 			$or: [{ triggers: { $exists: true } }, { actions: { $exists: true } }],
-	// 		}).fetch()
-	// 		for (const obj0 of objects) {
-	// 			const obj = obj0 as unknown as TriggeredActionsObj & TriggeredActionsOld
+	{
+		id: `Studios remove pre *WithOverrides properties`,
+		canBeRunAutomatically: true,
+		validate: () => {
+			const objects = Studios.find({
+				$or: [{ blueprintConfig: { $exists: true } }, { mappings: { $exists: true } }],
+			}).count()
+			if (objects > 0) {
+				return `object needs to be converted`
+			}
+			return false
+		},
+		migrate: () => {
+			const objects = Studios.find({
+				$or: [{ blueprintConfig: { $exists: true } }, { mappings: { $exists: true } }],
+			}).fetch()
+			for (const obj0 of objects) {
+				const obj = obj0 as unknown as DBStudio & StudioOld
+				Studios.update(obj._id, {
+					$unset: {
+						blueprintConfig: 1,
+						mappings: 1,
+					},
+				})
+			}
+		},
+	},
+	{
+		id: `ShowStyleVariants remove pre *withOverrides properties`,
+		canBeRunAutomatically: true,
+		validate: () => {
+			const objects = ShowStyleVariants.find({
+				blueprintConfig: { $exists: true },
+			}).count()
+			if (objects > 0) {
+				return `object needs to be converted`
+			}
+			return false
+		},
+		migrate: () => {
+			const objects = ShowStyleVariants.find({
+				blueprintConfig: { $exists: true },
+			}).fetch()
+			for (const obj0 of objects) {
+				const obj = obj0 as unknown as DBShowStyleVariant & ShowStyleVariantOld
+				ShowStyleVariants.update(obj._id, {
+					$unset: {
+						blueprintConfig: 1,
+					},
+				})
+			}
+		},
+	},
+	{
+		id: `ShowStyleBases remove pre *withOverrides properties`,
+		canBeRunAutomatically: true,
+		validate: () => {
+			const objects = ShowStyleBases.find({
+				$or: [
+					{ blueprintConfig: { $exists: true } },
+					{ sourceLayers: { $exists: true } },
+					{ outputLayers: { $exists: true } },
+				],
+			}).count()
+			if (objects > 0) {
+				return `object needs to be converted`
+			}
+			return false
+		},
+		migrate: () => {
+			const objects = ShowStyleBases.find({
+				$or: [
+					{ blueprintConfig: { $exists: true } },
+					{ sourceLayers: { $exists: true } },
+					{ outputLayers: { $exists: true } },
+				],
+			}).fetch()
+			for (const obj0 of objects) {
+				const obj = obj0 as unknown as DBShowStyleBase & ShowStyleBaseOld
+				ShowStyleBases.update(obj._id, {
+					$unset: {
+						blueprintConfig: 1,
+						sourceLayers: 1,
+						outputLayers: 1,
+					},
+				})
+			}
+		},
+	},
+	{
+		id: `TriggeredActions remove pre *withOverrides properties`,
+		canBeRunAutomatically: true,
+		validate: () => {
+			const objects = TriggeredActions.find({
+				$or: [{ triggers: { $exists: true } }, { actions: { $exists: true } }],
+			}).count()
+			if (objects > 0) {
+				return `object needs to be converted`
+			}
+			return false
+		},
+		migrate: () => {
+			const objects = TriggeredActions.find({
+				$or: [{ triggers: { $exists: true } }, { actions: { $exists: true } }],
+			}).fetch()
+			for (const obj0 of objects) {
+				const obj = obj0 as unknown as TriggeredActionsObj & TriggeredActionsOld
 
-	// 			TriggeredActions.update(obj._id, {
-	// 				$unset: {
-	// 					triggers: 1,
-	// 					actions: 1,
-	// 				},
-	// 			})
-	// 		}
-	// 	},
-	// },
+				TriggeredActions.update(obj._id, {
+					$unset: {
+						triggers: 1,
+						actions: 1,
+					},
+				})
+			}
+		},
+	},
 ])
