+++ meteor/server/publications/_publications.ts
@@ -5,6 +5,7 @@ import './expectedPackages'
 import './pieceContentStatusUI/publication'
 import './organization'
 import './peripheralDevice'
+import './peripheralDeviceForDevice'
 import './rundown'
 import './rundownPlaylist'
 import './segmentPartNotesUI/publication'
